requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6
numpy==1.25.2
python-dotenv==1.0.0
requests==2.31.0
